app/requirements.txt
amqp==5.3.1
anyio==4.10.0
beautifulsoup4==4.13.5
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
debugpy==1.8.16
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
kombu==5.5.4
packaging==25.0
playwright==1.55.0
playwright-stealth==2.0.0
prompt_toolkit==3.0.52
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==20.6
redis==6.4.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
